requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0
gunicorn>=21.2.0
